Validate for patientChart when creating/updating a chart; Check duplicate name of other object is wrong (Patient module).

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-patient/src/main/java/org/hisp/dhis/patient/action/patientattribute/ValidatePatientAttributeAction.java
@@ -96,7 +96,7 @@ public class ValidatePatientAttributeAction
 
         PatientAttribute match = patientAttributeService.getPatientAttributeByName( name );
 
-        if ( match != null && (id == null || match.getId() != id) )
+        if ( match != null && (id == null || match.getId() != id.intValue()) )
         {
             message = i18n.getString( "name_in_use" );
